refactor(tests): split test_manage_experiment.py into focused files
Split the 4331-line test_manage_experiment.py into four focused
sub-modules to stay within the Edit tool token limit:

- test_manage_experiment_cli.py — parser/CLI argument tests (5 classes)
- test_manage_experiment_repair.py — cmd_repair tests (2 classes)
- test_manage_experiment_visualize.py — cmd_visualize tests (3 classes)
- test_manage_experiment_run.py — cmd_run execution tests (30 classes)

No logic changes — purely mechanical split. All 125 tests preserved.

Closes #1164
